+++ src/Light.cpp
@@ -5,7 +5,7 @@ Light::Light()
 
 }
 
-Light::Light(Color color, Vector3d direction)
+Light::Light(Color color, Vector3f direction)
 :_color(color), _direction(direction)
 {
